Move secret versioning and snapshot functionality into ee and begin license scoping
