+++ backend/src/controllers/workspaceController.ts
@@ -8,7 +8,6 @@ import {
 	IntegrationAuth,
 	IUser,
 	ServiceToken,
-	SecretSnapshot
 } from '../models';
 import {
 	createWorkspace as create,
@@ -334,37 +333,5 @@ export const getWorkspaceServiceTokens = async (
 	
 	return res.status(200).send({
 		serviceTokens
-	});
-}
-
-/**
- * Return secret snapshots for workspace with id [workspaceId]
- * @param req 
- * @param res 
- */
- export const getWorkspaceSecretSnapshots = async (req: Request, res: Response) => {
-	let secretSnapshots;
-	try {
-		const { workspaceId } = req.params;
-
-		const offset: number = parseInt(req.query.offset as string);
-		const limit: number = parseInt(req.query.limit as string);
-		
-		secretSnapshots = await SecretSnapshot.find({
-			workspace: workspaceId
-		})
-		.skip(offset)
-		.limit(limit);
-
-	} catch (err) {
-		Sentry.setUser({ email: req.user.email });
-		Sentry.captureException(err);
-		return res.status(400).send({
-			message: 'Failed to get secret snapshots'
-		});
-	}
-	
-	return res.status(200).send({
-		secretSnapshots
 	});
 }
