+++ backend/src/ee/models/index.ts
@@ -1,0 +1,7 @@
+import SecretSnapshot, { ISecretSnapshot } from "./secretSnapshot";
+import SecretVersion, { ISecretVersion } from "./secretVersion";
+
+export {
+    SecretSnapshot,
+    SecretVersion
+}
