+++ backend/src/ee/routes/index.ts
@@ -1,0 +1,7 @@
+import secret from './secret';
+import workspace from './workspace';
+
+export {
+    secret,
+    workspace
+}
